test(tui): split landing draw env tests
Keep the landing draw implementation under the file-size gate.

Closes #582
